Separate data schema from models
