+++ api/models/member_job.py
@@ -1,6 +1,5 @@
 from api.models import Base
 from api.maap_database import db
-from marshmallow_sqlalchemy import SQLAlchemyAutoSchema
 
 
 class MemberJob(Base):
@@ -14,10 +13,3 @@ class MemberJob(Base):
 
     def __repr__(self):
         return "<MemberJob(id={self.id!r})>".format(self=self)
-
-
-class MemberJobSchema(SQLAlchemyAutoSchema):
-    class Meta:
-        model = MemberJob
-        include_fk = True
-        load_instance = True
