+++ api/models/pre_approved.py
@@ -1,6 +1,5 @@
 from api.models import Base
 from api.maap_database import db
-from marshmallow_sqlalchemy import SQLAlchemyAutoSchema
 
 
 class PreApproved(Base):
@@ -11,12 +10,5 @@ class PreApproved(Base):
 
     def __repr__(self):
         return "<PreApproved(email={self.email!r})>".format(self=self)
-
-
-class PreApprovedSchema(SQLAlchemyAutoSchema):
-    class Meta:
-        model = PreApproved
-        include_relationships = True
-        load_instance = True
