+++ api/schemas/member_session_schema.py
@@ -1,0 +1,10 @@
+from api.models.member_session import MemberSession
+from marshmallow_sqlalchemy import SQLAlchemyAutoSchema
+
+
+class MemberSessionSchema(SQLAlchemyAutoSchema):
+    class Meta:
+        model = MemberSession
+        include_fk = True
+        load_instance = True
+
